// src/app/shared/models/user.model.ts
export class User {
  _id: string;
  email: string;
  token: string;
  username: string;
  photoUrl: string;
  credentialProvider: string;
  firstname: string;
  secondname: string;
  birthday: Date;
  gender: string;
  ethnicity: string;
  country: string;
  current_year: Number;
  bannerUrl: string;
  bio: string;
  category: string;
  website: string;
  emailVerified: boolean;
}